requirements.txt
python-nmap~=0.7.1